refactor(idea_service): extract governance + showcase views (#163)
  before: 1895
  after:  1781 (-114)
